# docs/swagger.yaml
basePath: /v1
definitions:
  AudioShare_backend_internal_entity.Dump:
    properties:
      filename:
        type: string
      id:
        type: integer
      size:
        type: integer
    type: object
  AudioShare_backend_internal_entity.TrackFile:
    properties:
      bitrate:
        type: integer
      bucket:
        type: string
      content_type:
        type: string
      duration:
        type: number
      file_name:
        type: string
      id:
        type: integer
      object_key:
        type: string
      size:
        type: integer
      track_name:
        type: string
      uploaded_at:
        type: string
      uploader_id:
        type: integer
    type: object
  AudioShare_backend_internal_entity.User:
    properties:
      email:
        type: string
      id:
        type: integer
      login:
        type: string
      nickname:
        type: string
      password:
        type: string
      registered:
        type: string
      role_id:
        type: integer
    type: object
info:
  contact: {}
  description: API Для обмена аудио файлами
  title: AudioShare API
  version: "1.0"
paths:
  /auth/signin:
    post:
      consumes:
      - application/json
      description: Authenticate user and return tokens
      parameters:
      - description: User login credentials
        in: body
        name: request
        required: true
        schema:
          properties:
            email:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/AudioShare_backend_internal_entity.User'
      produces:
      - application/json
      responses:
        "200":
          description: User created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User registration
      tags:
      - auth
  /dump/create:
    post:
      consumes:
      - application/json
      description: Creates a new PostgreSQL database dump using pg_dump
      produces:
      - application/json
      responses:
        "200":
          description: Dump created successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create database dump
      tags:
      - dump
  /dump/list:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all database dumps
      produces:
      - application/json
      responses:
        "200":
          description: List of dumps
          schema:
            items:
              $ref: '#/definitions/AudioShare_backend_internal_entity.Dump'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all dumps
      tags:
      - dump
  /dump/restore:
    post:
      consumes:
      - application/json
      description: Restores a PostgreSQL database from a dump file using pg_restore
      parameters:
      - description: Dump file information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/AudioShare_backend_internal_entity.Dump'
      produces:
      - application/json
      responses:
        "200":
          description: Dump restored successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Restore database dump
      tags:
      - dump
  /tracks/download:
    get:
      consumes:
      - application/json
      description: Download an audio track file
      parameters:
      - description: Bucket name
        in: query
        name: bucket
        required: true
        type: string
      - description: Object key
        in: query
        name: objectKey
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Track file
          headers:
            Content-Disposition:
              description: attachment; filename=track.mp3
              type: string
            Content-Type:
              description: audio/mpeg
              type: string
          schema:
            type: file
        "400":
          description: Missing bucket or objectKey
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Download track
      tags:
      - tracks
  /tracks/info:
    get:
      consumes:
      - application/json
      description: Get metadata information about a track
      parameters:
      - description: Bucket name
        in: query
        name: bucket
        required: true
        type: string
      - description: Object key
        in: query
        name: objectKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Track information
          schema:
            $ref: '#/definitions/AudioShare_backend_internal_entity.TrackFile'
        "400":
          description: Missing bucket or objectKey
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get track info
      tags:
      - tracks
  /tracks/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload an audio track file
      parameters:
      - description: Audio file to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Track uploaded successfully
          schema:
            $ref: '#/definitions/AudioShare_backend_internal_entity.TrackFile'
        "400":
          description: Invalid input or file not provided
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: User not authorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Upload track
      tags:
      - tracks
  /users/:
    get:
      description: Get a list of all users (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            items:
              $ref: '#/definitions/AudioShare_backend_internal_entity.User'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - users
  /users/{id}:
    delete:
      description: Delete a user by ID (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: User deleted successfully
        "400":
          description: Invalid user ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete user by ID
      tags:
      - users
    get:
      description: Get user profile information by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            $ref: '#/definitions/AudioShare_backend_internal_entity.User'
        "400":
          description: Invalid user ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"